refactor(app-start) Make maps under one root component Closes #126
