+++ package.json
@@ -44,11 +44,12 @@
         "esri-leaflet": "2.5.1",
         "esri-leaflet-renderers": "2.0.6",
         "eventemitter3": "^4.0.7",
+        "html-react-parser": "^1.2.4",
         "i18next": "19.8.4",
         "i18next-http-backend": "1.0.21",
         "immutable": "^4.0.0-rc.12",
-        "lodash": "4.17.21",
         "leaflet": "1.7.1",
+        "lodash": "4.17.21",
         "proj4": "^2.7.0",
         "proj4leaflet": "1.0.2",
         "react": "^17.0.1",
